Assignment-1/requirements.txt
python-keystoneclient==3.15.0
python-swiftclient==3.4.0
flask==0.12.2
pyDes==2.0.1